requirements_enhanced.txt
# Core dependencies
customtkinter>=5.2.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
requests>=2.31.0
structlog>=23.1.0

# Enhanced dependencies for modern best practices
aiohttp>=3.8.0
aiofiles>=23.1.0
psutil>=5.9.0
pydantic>=2.0.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0

# Monitoring and logging
prometheus-client>=0.17.0
